Fixed an issue where tag changes weren't persisted..
If the change of tags causes no change in the length of the file, the file never gets an explicit (or implicit) flush. The file changes are not persisted on Android. This issue does not occur on Linux or Windows. This may be due to some low level file system caching problem, or a bug with Bionic / Android.

Manually calling flush after writing to the file is a simple workaround.